Don't broadcast ignored strings
Some setups might require specific tables or other kinds of queries
to not be run against a follower. For example, we have multiple
ActiveRecord::Base connections in our application for other databases
and we don't want queries from those connections run against our
primary follower as the tables don't exist.

This change allows a user to pass in a regular expression that will
be ignored in the Broadcaster:

    Burninator.new.broadcast(:ignore => /blogs/)

Any queries that match this regular expression will not be broadcast to
the warming process.

* Bump to 0.6.0
* Move optional parameters from Burninator constructor down into #broadcast
* Update documentation to reflect exception if the environment is not
  properly setup